refacto: move functions and const to helper/
